backend/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.30.1
SQLAlchemy[asyncio]==2.0.30
asyncpg==0.29.0
alembic==1.13.2
python-jose==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==3.2.2
pydantic==2.7.1
pydantic-settings==2.2.1
docker==7.1.0
celery==5.4.0
redis==5.0.4
python-multipart==0.0.9
email-validator==2.1.1
tenacity==8.3.0
